+++ app/src/main/java/com/dqr/www/multitypestudy/mine/bean/MineNavBean.java
@@ -4,21 +4,21 @@ package com.dqr.www.multitypestudy.mine.bean;
 import com.dqr.www.multitypestudy.mine.MineNavigateType;
 
 public class MineNavBean {
-	private byte type; //类型
+	private int type; //类型
 	private int iconId;
 	private String iconName;
 
-	public MineNavBean(@MineNavigateType.Tab byte type, int iconId, String iconName) {
+	public MineNavBean(@MineNavigateType.Tab int type, int iconId, String iconName) {
 		super();
 		this.type=type;
 		this.iconId = iconId;
 		this.iconName = iconName;
 	}
-	public byte getType() {
+	public int getType() {
 		return type;
 	}
 
-	public void setType(byte type) {
+	public void setType(int type) {
 		this.type = type;
 	}
